+++ vscode/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the Karate Debug extension will be documented in this fil
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.6] - 2026-03-09
+
+### Added
+- One-time notification for existing users about price reduction from $29.99 to $9.99
+
 ## [0.7.5] - 2026-03-09
 
 ### Fixed